refactor(webapp): rename component EndpointsList to EndpointsTextList
- rename the component
- build an string with the list
