+++ webapp/src/components/EndpointsTextList/styles.js
@@ -1,5 +1,6 @@
 export default (theme) => ({
   container: {
     padding: theme.spacing(2),
+    whiteSpace: 'pre-line',
   },
 })
